refactor: get protocol 삭제
